+++ frontend/src/components/header/Header.js
@@ -1,9 +1,4 @@
-import React, {Component} from 'react';
-import {BrowserRouter as Router, Link} from 'react-router-dom';
-import Sidebar from './Sidebar';
-import ChatArea from './ChatArea';
-import MessageArea from './MessageArea';
-import WebSocket from "./WebSocket";
+import React from "react";
 
 class Header extends React.Component {
     render() {
@@ -77,92 +72,4 @@ class Header extends React.Component {
     }
 }
 
-class Divider extends React.Component {
-    render() {
-        var lineStyle = {
-            marginTop: '70px',
-            display: 'block',
-            border: 0,
-            borderTop: '1px solid #E2E2E2',
-            padding: 0
-        };
-        return (
-            <hr style={lineStyle}></hr>
-        );
-    }
-}
-
-class Footer extends React.Component {
-    render() {
-        return (
-            <div className="footer">
-                <span>최연정 님이 입력 중</span>
-            </div>
-        );
-    }
-}
-
-class Root extends React.Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            chatAreas: [],
-            chatAreaRefs: []
-        };
-        this.addItem = this.addItem.bind(this);
-        this.sendMessage = this.sendMessage.bind(this);
-        this.printMessage = this.printMessage.bind(this);
-    }
-
-    addItem(cId) {
-        let reference;
-        this.setState((prevState) => {
-            return {
-                chatAreas: prevState.chatAreas.concat(<ChatArea key={Date.now() + cId} cId={cId}
-                                                                ref={chatArea => reference = chatArea}/>)
-            }
-        });
-        this.setState((prevState) => {
-            return {
-                chatAreaRefs: prevState.chatAreaRefs.concat(reference)
-            }
-        });
-    }
-
-    sendMessage(cId, text) {
-        this.sock.sendChatMsg(cId, text);
-    }
-
-    printMessage(message) {
-        for (let ref of this.state.chatAreaRefs) {
-            if (ref.props.cId == message.channelId) {
-                ref.addItem(message);
-            }
-        }
-    }
-    render() {
-        return (
-            <div>
-                <Sidebar addChatArea={this.addItem}/>
-                <Header name="general"/>
-                <Divider/>
-                {this.state.chatAreas}
-                <MessageArea sendMessage={this.sendMessage}/>
-                <Footer/>
-                <WebSocket ref={sock => this.sock = sock} printMessage={this.printMessage}/>
-            </div>
-        );
-    }
-}
-
-class App extends Component {
-    render() {
-        return (
-            <Router>
-                <Root/>
-            </Router>
-        );
-    }
-}
-
-export default App;
+export default Header;
