+++ src/main/resources/templates/channels/channels.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>Channels</title>
-    <link href="/public/static/css/main.fcb3c79f.css" rel="stylesheet"/>
+    <link href="/public/static/css/main.2b962602.css" rel="stylesheet"/>
     <!--<link rel="stylesheet" th:href="@{/webjars/bootstrap/4.1.0/css/bootstrap.min.css}"/>-->
     <!--<script th:src="@{/webjars/sockjs-client/1.0.2/sockjs.js}"></script>-->
     <!--<script th:src="@{/webjars/jquery/3.3.1/jquery.min.js}"></script>-->
@@ -12,14 +12,14 @@
     <!--<script src="/js/popup.js"></script>-->
     <!--<script src="/js/file.js"></script>-->
     <script src="/js/notification.js"></script>
-<script>
-    document.addEventListener('DOMContentLoaded', function () {
-        Notification.requestPermission();
-    });
-</script>
+    <script>
+        document.addEventListener('DOMContentLoaded', function () {
+            Notification.requestPermission();
+        });
+    </script>
 </head>
 <body>
 <div id="root"></div>
-<script type="text/javascript" src="/public/static/js/main.ed6e4807.js"></script>
+<script type="text/javascript" src="/public/static/js/main.28774a8c.js"></script>
 </body>
 </html>
